requirements.txt
# ── Core ML / CV ────────────────────────────────────────────────────────────
deepface==0.0.93
tensorflow==2.15.0          # required by deepface
tf-keras==2.15.0
opencv-python==4.9.0.80
scikit-learn==1.4.2
numpy==1.26.4

# ── Image I/O ────────────────────────────────────────────────────────────────
Pillow==10.3.0
pillow-heif==0.16.0         # HEIC/HEIF support

# ── Data / Visualisation ─────────────────────────────────────────────────────
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2

# ── Google Cloud Storage ──────────────────────────────────────────────────────
google-cloud-storage==2.16.0
google-auth==2.29.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9     # needed for file uploads
slowapi==0.1.9              # rate limiting

# ── Config / Env ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0               # for FastAPI test client


## Test